Ingore commits with over 50 files changed
This is temporary and will probably be removed soon. I'm assuming that
if there are over 50 files involved in a commit, we just vendored some
library or something. I might make this do something like "cp
foldername ..." instead of just ignoring it. I mostly want to avoid
typing out the ace editor stuff when you run this on itself.